7.45 ANN regression model
Fixed graphs in previous experiment.

Made the first neural network of this course! Employed Stochastic Gradient Descent to train a model on a random set of data. Gave the model x values, and had it predict the y values.

Shows how to create the model, model layers, learning parameters, training, and plotting the results.