7.52 Multiple Output ANN
Used an ANN to categorize data points in the well known iris dataset.

The model outputted three values, which correspond to the probability of a flower belonging to each category.

The softmax function was used to select the largest of these 3 probabilities, and categorize the flower to the (hopefully) correct category.

Used the CrossEntropyLoss() function instead of the other loss functions used in previous lessons. This is used when there is more than one category.

Results were graphed like normal, with additional explorations of graphing the raw output of the model to explore relationships to model performance.